requirements.txt
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
mysqlclient==1.3.13
pytz==2018.9
sqlparse==0.3.0
